third_party/WebKit/LayoutTests/typedcssom/inlinestyle/transform-perspective-expected.txt

This is a testharness.js-based test.
FAIL Perspective read from a StyleMap is correct assert_equals: expected "perspective(0px)" but got ""
FAIL Set Perspective into the inline StylePropertyMap Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
Harness: the test ran to completion.
